+++ app/controllers/api/v1/fibonacci_controller.rb
@@ -1,5 +1,10 @@
 class Api::V1::FibonacciController < Api::BaseController
 
+  def index
+    @fibonaccis = Fibonacci.last(10)
+    render json: @fibonaccis
+  end
+
   def create
     unless permitted_fibonacci[:n]
       render json: { error: "Please set the param 'n'" }, status: :bad_request and return