+++ db/migrate/20221021171427_create_fibonaccis.rb
@@ -4,7 +4,7 @@ class CreateFibonaccis < ActiveRecord::Migration[7.0]
       t.integer :value
       t.integer :result, limit: 8
       t.decimal :runtime
-      t.time :created_at
+      t.datetime :created_at
     end
   end
 end