+++ spec/requests/api/v1/fibonacci_spec.rb
@@ -2,7 +2,16 @@ require 'rails_helper'
 
 RSpec.describe "Api::V1::Fibonacci", type: :request do
 
-  describe "POST /create" do
+  describe "GET /fibonacci" do
+    subject { get "/api/v1/fibonacci" }
+
+    it do
+      subject
+      expect(response).to have_http_status(:success)
+    end
+  end
+
+  describe "POST /fibonacci" do
     subject { post "/api/v1/fibonacci", params: }
 
     context "when parameters are present" do